Allow key-value pairs with value `null`
For some full instantiations of target properties, key-value
pairs with value null are automatically derived (for example, the
zephyr platform). When trying to generate distributed/federated
programs, the system needs to generate lf programs and thus
convert the internal representation of target properties to text.
This commit fixes a NPE during string-building if a target property
(port, in case of the platform properties) has a null-value (i.e.,
explicitly set to null, or unspecified).