Fix LF property generation of empty key-values
Now, platform properties with value ´null` are not output
whatsoever when synthesizing LF programs. For example,
when synthesizing federated zephyr programs, the `portValue`
might not be set (i.e., `null`), and before, the program
would generate the `portValue: null` property. While not a
problem in this case, his fixes a rather unrelated unit test
fail.